requirements.txt
playwright>=1.40.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.92.0
